TallyCounter Update 1 6/13/2018
Tally Counter is an app for counting anything and everything. It's
multi-counter functionality allows you to keep track of all the
important things in your life that are countable.
Current Features:
Create, edit, reset and delete Multiple counters
Customize the color of each counter
Supports negative and decimal numbers
Events log, a list of changes made to a counter
Free, no Ads